Resolve "Update the DRC interface to exploit recent property API features and simplify terminology"

Closes #935 and #1710

See merge request opencpi/opencpi!421

+++ runtime/hdl/include/HdlWciControl.h
@@ -84,7 +84,7 @@ namespace OCPI {
 	}
 	return offset & (OCCP_WORKER_CONFIG_SIZE - 1);
       }
-      void throwPropertyReadError(uint32_t status, uint32_t offset, size_t n, uint64_t val) const;
+      void throwPropertyReadError(uint32_t status, size_t offset, size_t n, uint64_t val) const;
       void throwPropertyWriteError(uint32_t status) const;
       void throwPropertySequenceError() const;
